Created Sound Class Mix so can change the audio levels with the options menus. Updated options wipe animation in pause menu options widget. Storing sound volume values in game mode. Setup sliders functions in options menus. Unpausing whilst in options menu functionality implemented. Persistence between options menus when within game for audio and look sensitivity.
Created sound class mix which has got all of the sound classes in and will control the volume of all of the sound classes in the game.
In the options menu and the pause options menu we now properly set the sound mix class override and then push the sound mix modifier for the sound class mix to properly change the volume based on the value of the slider.

Options pause menu wipe now affects the return button.

New array of floats in the game mode and new function to update this array in the game mode. This is the volume levels for the different sound classes so can use them as persistence floats between the options menu and the pause options menu and to ensure they remain the same when opening and closing the menus.

New events in the options and pause options widgets for updating the sliders and values of the sliders and text with the volume levels and the look sensitivity.

When unpausing with P whilst in the pause options menu now properly works with wipes and everything so yay.