I Doubt It Doubt Check Update. "Fixed" sorting bug. Unique card indicator. Rules widget interface and implemented 2 events to consolidate enlarge and shrink rules functionality. Rules popping up after alt tabbing fixed. Removed holo card rep from auction adding cards.
I doubt it doubt check now ignores unique cards since they shouldn't be checked since some of the effects of the unique cards are very obvious that the person has just lied so also need to update the rules to let the player know about this.

Sacrificed player being able to look whilst sorting their cards so that the cards sort into the correct positions and rotations and don't mess up as much whilst moving camera whilst sorting.

New widget for unique card indicator with text so not showing holo version of the unique card but instead showing that the specified unique cards effects have been activated. The functionality for this is currently only in the I doubt it game manager but could probably also be useful in the auction manager. 4 Spaces currently in the widget for the cards since its designed for IDI. Setup function in the widget. Setup functions for position and visibility in the IDI game manager. Update function in IDI game manager called on end round passing in all cards played to loop through them to see which are unique cards and which ones should be shown since we don't want all unique cards to be shown since some are supposed to be hidden until doubted so there is an array of unique card names as local variable array to check against for what cards shouldnt be shown. New event in the IDI game manager to fade this widget as well.

Created an interface specifically for the rules widgets can probably just change this to be all widgets so don't need to cast will probably go through all widget code tomorrow and update everything to more closely use polymorphism and to improve optimisation generally and for the code base.

When the rules finishes shrinking now it removes itself from the parent so removed the alt tabbing bug where the rules would show up again with the mouse cursor after alt tabbing.

Unticked the holo representation from all add card to piles in auction for the played cards and won cards since the zoom in mechanic kinda replaces this now so auction should only have holo rep for the auctioning card.