New Enum for NPC States. Game over function in game mode. New actor and new enum for positioning. MOVING CHARACTERS LOGIC AND FUNCTIONALITY CHANGED AND UPDATED. More level organisation.
New enum for setting up NPC states currently only 4: idle, character specific action, dialogue, and in-game.

Game over function created in game mode.
Move characters after game over function created in game mode as well.
Calling game over function for game mode in all of the game managers in their respective game over functions. Also setting all the widgets for the game in the game over function not visible in this function but might change this in the future.

Game over game mode function:
First updates the maps for games completed and games won or mini games won and mini games difficulties won.
Then starts a camera fade for 4 seconds holding when finished and then calls the move characters after game over with a 4 second timer so it does it when the camera fade reaches black.

Move characters after game over function in game mode:
Moves all characters to their neutral positions using Move NPC to Neutral macro which needs an NPC actor passed to it. Using the map of neutral positions stored in the game mode which is setup in the setup character neutral positions macro. This macro finds the character position actors that have intended game as none and adds them to a map using the NPC actor as the key.
Then we more the player to their correct position using a character position actor that has the corresponding tag for the game they have just completed with the tag being the game name shortened followed by _Finished.
Then finally we change the input mapping contexts to be bar instead of the game one. Also updating button prompt shown, setting current playing game to none and setting game active to false and then starting a camera fade to clear over 2 seconds.

New function in IDI game manager for clearing all player hands and adding them to the main pile. This is called with a 4 second timer in the game over function in the IDI game manager and is used to clear all of the player's hands so that they are then not walking around with their cards still in their floating splines. Will probably need to do this for auction but will first test it for IDI and implement it later for Auction.

Organised the level more with renaming and putting stuff into folders.

Only thing missing from the end game functions now is some polish and the win declaration widget so that the player actually knows if they have lost or not.
New functions in all the main game managers for hiding all the game widgets including the current turn indicator in the turn manager for all games except auction. Will probably need to add some functionality for these widgets being set back to visible again on game setup instead of on begin play.