Clearing hovered card on card play. Player can no longer play cards whilst doubting is being showcased in IDI. Sorting spamming bug fixed. IDI and Auction no played cards checks.
New function in the current hand component for clearing the hovered card. This is called in the auction and IDI game managers when we play a card(s). This function sets the hovered card to be null and then calls the hover first card function in the current hand component with a 0.25 timer to account for cards moving.

Famine update in IDI to clear hovered and selected cards when swapping so there's no bugs need to properly test.

New bool in the player for being able to sort their cards.
This fixes a bug with spamming the sort cards button which would break the ignore look input commands and would cause the player to be unable to look around.
Increased the speed of the card sorting to account for this so the cards move with 0.125 seconds instead of 0.25 now. So can still spam it basically but it fixed this bug.
Also added some extra checks for sorting so can only do it whilst the game is active and whilst we are in auction or IDI just in case.

New check for end turn in the player specifically for IDI. We cannot end our turn whilst doubt result is being displayed. This is controlled by a new bool in the IDI game manager.
New check in end round function for auction to check if the current looping player actually has any selected cards and if they don't then we just ignore the loop and move on to the next character. Haven't tested this but I think this should work.