File organisation. Deleted unused first person blueprints and animation blueprint and meshes.
Basically got rid of the standard first person template files that are unused since we are using basically no mesh for the player and we don't need a gun and projectiles etc.
Moved the main map into the maps folder and changed its name so easier to access.
Changed the standard maps for editor loading and game loading to the test level and the main level.
Changed the game mode for the main level to be the game mode we use for functionality.