Door system.
New actor for the bar door which will replace the current doors that are on the bar but haven't done that yet.

This is an event driven door interaction system detecting the player entering and exiting.
Opens in the direction in which the player is moving using 2 collisions one for entering the bar and one for exiting the bar.
Closes when the player exits both collisions.
If the player stays within one of the collisions whilst the door is open then the door will remain open.
Modular time to close after exiting.
Modular time to open using play rate of timelines.
Smooth rotation of doors for both opening and closing using timelines with float curves.
Using relative rotation of door static meshes to rotate them to the intended points.

Modular angle to open door using rotation change float variable that is exposed as well.

Using timers to delay certain actions as well as checking if the player is still overlapping with the actor after some delays after certain actions like leaving one of the collisions.