Main menu "cutscene". New input mapping context specifically for main menu.
New mapping context specifically for the main menu. This contains click, look, exit credits and pause input actions.
Bar mapping context now no longer contains the exit credits input.
New functions in the game mode for entering the main menu and exiting which just changes around the current mapping context. Enter main menu is called when we overlap with the main menu collision and pass the do once.
Exit main menu is called when we finish the timeline in the to cutscene graph.

New functionality in the to cutscene graph.
Also new variable for the amount of time for the timeline in the cutscene graph.
New timeline in the cutscene graph which has a float curve used to lerp along the new "cutscene" spline which is basically just moving the player along it. Not currently using the rotation and instead allowing the player to look around whilst on it but this can always change.
On finished for the timeline we call leave main menu controls in the game mode, to main prompts in the bar button prompt widget and finally call reset move to main which would the allow the player to enter the main menu again if they were to enter the collision again.