Didn't Save Main Menu. NPC mesh choosing. NPC update interaction after game over. Game mode checking if all games have been played in game over function. Entire game completed function in game mode. Temporary game completed widget.
NPC in construction script now changes its skeletal mesh asset based on the given NPC name that the NPC has.

NPCs now have their interaction disabled when the game that they are assigned has been completed so that the player cannot play against them again until the level is reloaded.
This is done through 2 functions in the NPC. One of which is unused as of now which is to turn the interaction back on.
The one to turn interaction off basically sets the interact collision to no longer generate overlap events and then does the functionality for when the player leaves the interaction collision.
There is a new function in the game mode to call this disable interaction function which converts the current playing game enum to the main games enum and then gets all NPCs and loops through them until it finds the NPC that started the currently playing game since this is called before we reset that variable. We then store this NPC as a local variable and stop the loop and then call the disable game start interaction in the found NPC.
This function is called in the move characters after game over function but now before we reset all player mapping contexts and the current playing game.

Also new functions in the game mode for checking if we have completed all of the 3 main games. This just loops through the map of games to completed bools and checks if any of them are false and if so then we store a local bool as true and then return this inverted since if it is true then that means we have not played all the games. This function is called in the game over function in the game mode before we call the move characters after game with a timer we instead now branch and if this is false we do that if it is true then we call a new function called Entire Game Completed with a 4 second timer.
This other new function currently just creates a widget of new class game completed and adds it to the viewport then setting the player input to UI only and showing mouse cursor.
This new widget is a temporary one that just says game completed with 2 buttons for restarting or quitting the game and functionality for these buttons.

Need to test these functions but will do later since want to try importing rigged model.