Reimported the base third person template content and then deleted all the third person stuff since we don't need it just need the UI skeletons. Imported rigged version of Dante successfully.
As a note for importing the models make sure to do the following:
set the skeleton as the ue5 skeleton.
change the vertex colour import options to override.
set the normal import method to import normal and tangent so that the smoothing data is used correctly for the mesh.
Changed all the NPCs to use this mesh for their selection for now and they are currently just using the Manny animation blueprint but I'm going to start working on animation functionality soon so that will change.