More animation anim graph setup. Update enum NPC states.
State machines setup with idle animations instead of actual animations for now but logic is mainly there just need counts for how many animations of certain ones they have for the blend poses by int.

On enter state animation events to calculate randomly which animation should play.

Functionality setup in the NPC actor itself for determining if it should enter the character specific state machine. This needs finialising with animation notifies to turn the check on and off.

New bools in NPC for use in animation blueprint: in game ( which needs functionality for setting when the NPC enters a game), my turn? which acts the same way as the bool in the player determining if it is their turn this will also need tweaking so that it stays active whilst the animation plays for certain games like darts, should do character specific action as described above.

Threadsafe function for updating the state done for now in the animation blueprint.

Currently for dialogue just layered blend per bone with state machine for dialogue being head up and the normal state being head below.
Dialogue state machine is the only one not properly setup with logic yet.