Door frame is now using complex collision instead of simple so player can enter bar. Updated door actor to match doors in the bar. Merged the jukebox meshes to create one for use in the jukebox actor. Jukebox actor replaced jukebox static meshes. Darts bug fixing. Dart board actor replace dart board static mesh.
The door frame was originally using simple collision which was blocking the player from entering the bar so this is now changed to be using the complex instead.

Altered the scaling of the doors in the automatic door blueprint to more closely match the doors that were originally setup for the bar and replaced those doors with the automatic door blueprint so that the player can enter and exit the bar now.

Jukebox meshes merged into one.
Jukebox actor updated with this mesh.
Jukebox replaced with jukebox actor.
Rearranged jukebox components.
Changed colour of widget component for jukebox interaction since the post process makes them really bright. This is going to be the same for all diegetic UI probably.

Darts for some reason was messing up in the main level.
Finally figured out that the directions on the dart board needed swapping around FOR SOME UNKNOWN REASON THAT WAS REALLY FUN TO FIGURE OUT ACTUALLY.
This is fixed now and darts works as intended in the main level.
Replaced the original dart board with the dart board actor.

Outside looping functionality updated to select randomly from a set of available teleport points instead of one since there are 3 paths that the player could take and it would be more interesting and disorientating to have them be randomly teleported when entering the loop collision.
Put all the loop collisions and points into the main level.

Will move more stuff tomorrow.