Tweaked door collision sizes. Changed sizes of UI for darts starting interaction. Changed positioning of the UI for the dart board in the level. Increased z of camera so player is looking a little higher. Moved the spline for card hand down a little after moving camera up. Main menu setup in level.
Changed the sizes of the door collision for the door actor in the main level.
Changed the sizes of the text in the various dart associated starting widgets.
Changed the positioning for the widgets for the dart board in the level.

Increased the z position of the player's camera and then slightly lowered the current hand spline since the camera is the parent of the spline.

Moved the main menu into a relatively correct position for the main level as well as the components for the main menu.
Need to add some more functionality for the main menu since it gets a little annoying leaving the bar and being immediately forced into the main menu and not being able to explore so maybe need some sort of toggle or interaction that the player chooses rather than automatic collision detection after the first time of the main menu being used.