Main Menu Auto detection activation and deactivation.
New functions in the main menu actor for enabling and disabling the auto detection which just sets the generate overlap events true or false.

New macro in the door actor for changing the auto detection of the main menu. This calls the deactivate auto detection and then sets a timer for the activate auto detection function and sets the timer. This allows the player to use the door out of the bar and not be automatically put into the main menu until the timer runs out.