NPC turns head to look at player now.
Update in both the NPC animation blueprint and the Manny animation blueprint
In event graph set player reference after 0.2 second delay on blueprint animation initialisation and we then use this player reference to get the actor eyes position and subtract about 10 and then set this as a new variable in the player reference is valid in the event graph.
We then use the look at skeleton node in the anim graph using the head bone with the look at axis at 0,0,1 and a look at clamp 55 and interpolation at 1.
We must convert to local component and then back again for the look at skeleton node to work.

This now means that the NPC will look at the player when the player is within a certain angle of the NPC.