Pool Mini Game Progress
Big push here so --- Stuff I reused:

Start Game Diagetic UI and functionality (interact collision boxes that trigger a trace that triggers the UI appearing, click on button and game starts)

Pool Game:

Pool Table actor - multiple empty static meshes that are placed in the locations the balls will be teleported to when the game begins. This is done with a map of a pool ball name enumerator with the values of the empty meshes being set on begin play;
Togglable overhead camera that shows the entire table.

Pool Ball Actor -

A static mesh of the ball

Cue Ball Actor - A child of the pool ball actor. Has a camera attatched to it that can be rotated and aimed with the A and D keys. Whichever way the camera is facing is the way the ball will projectile towards when the Hit Ball function is called. Gets the forward vector of the camera and adds impulse in that direction to do this.

Pool Game Manager -  ALL OF THIS HAS NOW BEEN EFFECTIVELY DELETED BY GIT HUB WILL RE DO NEXT PUSH :))

The game manager spawns all the balls on begin play and creates an array of them.

"Setup Game" Hides UI and fades the camera out before setting the ball positions on the table. Then calls the start game function.

"Start Game" as of right now simulates the balls physics and does some setup for which character's turn it is.

Believe thats it.