NPC head tracking when entering interact collision. NPC tracking different characters.
New enum in the NPC for character to track.
New bool for should head track in NPC.
New timer stored for disabling head tracking when leaving interact collision.
Enabling and disabling head tracking for entering and exiting interact collision with NPC.

3 new functions in NPC interface for enabling and disabling head tracking and one for updating the character to look at passing the character name num. Enable head tracking function also passes the character to look at.

Same variables in the animation blueprint.
Functionality in the threadsafe update animation blueprint to update should head track and the character name to track using property access.

Functionality in the event graph for getting the target position to look at using the character name to look at and checking if it is valid before using the actor reference.