Leading cat functionality started. NEW CARD MATERIALS IN AND WORKING OH YEA. Updated card set material and holo card set material functionality. Setting card deck via card dealer deal cards function. New parameter in update holo card info function for what card deck we are using.
Basically just did all the functionality for the leading cat obviously not fully done since need to do some edge case implementation as well as polishing but yea.
Obviously also its not really got any mesh right now but yea.
Need to setup a new animation blueprint specifically for the leading cat relative to the schrodinger's cat.
Setup a simple bool in the leading cat actor for when it is moving which doesn't account for if the timeline is active but not moving the cat.

JESS CARD MATERIALS AND TEXTURES IN NOW AWESOME!

New function in the card interface for updating the deck material for the card passing an integer.
This function is called in the card dealer when dealing the cards passing a random integer between 0 and 6.

When setting holo card rep visible we call the update holo card info passing the deck int from the first card in the pile since they will all be the same and we know that there will be at least one card in the pile for the holo rep to show.

The only thing ive noticed with the cards is that the holo card cannot be transparent without using another material or changing the material which messes it up for the other cards and also that the cards can be really dark so there may need to be some specific lighting for the cards themselves especially in the main level.