Dialogue Advancement. Jukebox track title widget and functionality. Updated method of storing songs in jukebox for easier track title display. New finished songs added for jukebox.
Do Dialogue interface function.
Stop dialogue interface function.
New enum for dialogue contexts.

Threadsafe updating dialogue info: in dialogue, dialogue to talk to and dialogue context.

New widget for displaying the track name of the currently playing song on the jukebox.
New enum for the song names that will be playing on the jukebox.
Changed the array of sound cues to a map of sound cues to names of songs.
Changed functionality in jukebox to accommodate this change to map rather than array.
Fixed a bug with going next song when at the end so it loops back around as it should now.
New functionality in jukebox for showing and hiding this new widget for track title.

New functionality in jukebox for setting jukebox track in widget when showing title and when changing title.
Functionality in jukebox and widget for playing a springy animation when pausing and resuming the song so some feedback other than audio feedback.