Dialogue struct and data table. Further implementation setup for dialogue selecting audio and anim graph.
New struct for dialogue which is just an array of possible sound cues to play.
New data table for this dialogue struct.
The row name of the data table will be in the following format: Name of NPC followed by Name of Recipient of Dialogue followed by Context of Dialogue.
Functionality in the NPC for getting the array of sound cues for the relevant row in the data table converting the enums to strings and appending them.
Then getting a random sound cue out of the array for use in the audio component on the NPC for the dialogue. Also setting the index of the sound cue as a variable. Using this in the animation blueprint so also doing property access in threadsafe update to get this. Using this integer to blend by bool in the dialogue state machine states.