Fixed NPC card hand rotation. Games setup in main level. Tweaked player current hand spline. Auction card pile specific child blueprint. Functionality for correctly choosing which card pile for auction. Fixed rat screw add card to bottom pile rotation. Lots of changes and bug fixes and general improvements to all 3 major games.
Current hand component update visible hand fixed for NPC selecting the bone to look at from card position to use the new bone for the NPC skeletons.

Spent a lot of time setting up the games in the main level.
Spent a lot of time tweaking different things like positioning of players in games, positioning of card piles, fixing issues with card piles, changing the locations of the current hand splines in NPCs and players.
Spent a lot of time getting the end game functions in the main games working without really any bugs which is good. Probably still some edge cases to find.
Ending the game completely now properly works.
Changes in the game mode.
Restarting the level through game completed doesn't fully work so that needs a bit more work to fix it.
Need to add in the scaling difficulties for the main games after completing one of them we increase the difficulty a bit for the other games.

New actor for auction specific card piles which is the same but contains an int for the player that it is associated with: player being 0. This is used in the auction game manager to correctly add cards played and cards won to the correct pile since we need modularity with this whilst changing around the player positions for the games.

Fixes and more functionality for player collisions whilst in the games and resetting this when game finishes.
More functionality for removing all of the cards on game finish and then using them again in another game.

Rat screw zoom button actually set now since it wasn't before.