Progress on ball potting + scoring functionality
BP Pool Pocket has a sphere collision that only overlaps with pool balls. When overlapped checks if they potted black, which will end the game. if not, adds the ball potted to an array in the game manager.

"Pot Ball" function (name gonna change for this because its just wrong) will loop through all of the potted balls checking for if the cue ball was potted and if the balls potted belong to the active player. Fouls are set as a bool and the same if the player scored. Once loop is completed the next turn function will be called, and all of these variables will be inputted.

Next turn checks if there is a foul, and if the active player is playing again, and sets the active player accordingly.

Win game function has input bool for if the active player wins and prints who ever wins currently. obviously needs work

Likely will have to track whether the balls are moving and once they have all stopped moving then call the "pot ball" function. wat should i name it 2 :P