NPC eye tracking functionality implemented.
NPC animation blueprint 2 more look at skeletal nodes for the eye joints to follow the player.
Can turn this on and off with just connecting and disconnecting these nodes, using new vector for the eye look at location setup in the same place in the event graph as the main look at location.
Will try with blinking animation later when I get that.