Tuna model added to game and schrodinger and leading cat actors.
Tuna model added to the game in a shared cat assets folder in misc mechanics folder.
Everything seems good with it idk using the same settings as the other models for smoothness.
Changed the skeletal mesh models for the leading cat and schrodingers cat actors to this new tuna model.

Also found how to circumvent the streaming pool restrictions for textures.
Command: r.Streaming.Poolsize and then an amount relative to how many gbs your GPU has I had to use 8000mb for it to load all textures.