NPC scalar difficulty with game completion.
New function in the game mode for changing the difficulty of after completing a game this function is called in the game over function after checking if we have completed all the games and this is false.
This function loops through all the NPC and possibly increase their difficulties for IDI and auction by 1 depending on a bool weighted with 0.75 with rat screw always increasing by 1. If the increased byte is greater than the maximum number of entries - 1 then we set the difficulty to expert instead since that is the maximum.