Auction advanced end round functionality and won card indicators.
Auction delaying moving the won card to the corresponding pile with setting a timer for the function in the end round function. This function is called Finish Non-Draw Round.
New functionality for won card counters using the card counter and functionality from I doubt it game manager with passing the amount of cards as the amount of won cards for the player that has just won the round.
The update won card counter function is called in the end round function.

New functionality for unique card indicators for auction.
Setup and update functions for the unique card indicator on the auctioning pile.

Think of the lagging causes is the main menu when the widgets are visible and not interacting with them so setting them not visible seems to fix this a bit.