Changed a lot of components hidden in game. Edited some widgets tint and colours for the main level. Changed jukebox widget. Dialogue notify class. Fixed doubting own turn. Player current hand spline actually centered now. Cards now use all lighting channels. Glitchy fix for card movement. Fixed next and previous card timings. Changed the IDI fade out timeline lengths.
Lots of things that should be hidden in game now actually are.

Edited the tint and colour on some of the widgets since they were way too bright in the level.

Changed the jukebox title widget animations and setup so the text is not blurry anymore using the canvas panel scaling method with the widget component being scaled smaller.

New notify class specifically for dialogue notifies which will be used for dialogue strands of repeated lines one after another between characters. This notify will have functionality within it to call do dialogue passing its stored context and speaker and intended recipient so can still use that system. Don't know why we never used these before in previous projects since these could've made things so much easier since it is just notifies but with unique classes that we can create. Should really consider doing this for a lot of things that we come across where we can't do exactly what we want with the already given standard classes since that is actually programming.

Fixed a logic error with doubting so now using a or instead of a and for the first check since we could actually doubt our own played cards with an and bool.

The current hand spline didn't look centered eventhough according to the vectors it should've been so slightly tweaked its positioning in the player and NPC to fix this.

Cards now actually use all of the lighting channels so can more easily see them in the main level.

Now when a card finishes its smooth card movement timeline and after passing the player camera validity check we then check if the current position of the card is the same as the card's intended position in the current hand spline of the player since if the camera is valid then that means that we are adding to the player's spline.

Fixed issue with the next and previous card hold down feature in the player using different values now so more smoothly from testing but may need to keep being changed idk.

Changed the length of the IDI timelines for fading out widgets to be longer so can see them in their glory for longer.