Changed unique card indicator for IDI to be properly scaled and not blurry. More widget changes functionally and aesthetically for the main games. Rat text pop up widget new functionality. Changed zoom in positions/rotations for all main games. More dart widget changes.
Unique card indicator is now scaled properly and no longer blurry. Also now has a background as well.
Draw widget changed.
Player cards round IDI changed.
Current round card value widget changed.
Jukebox title widget changed.

Rat screw pop up text more functionality for it showing the slap asset in both the widget itself and in the rat screw game manager. Now world space and not screen space like other UI in the game.

Altered the rotation and positions of the zoom in points for all of the main games.

Changed more for the dart widgets. Mainly aesthetically and also changing the animations to be on the overlays rather than just the buttons as well as changing the animations so that the scale doesn't exceed the shown and so there is no clipping visible.
Thought darts had broke again but just needed to compile the dart game manager again and it started working again so that's not worrying at all.