Button prompt use new assets. Removed move from button prompt since pretty self explanatory. Main Menu backtracking volume disabling on first entry.
Button prompt widgets updated to use all the new assets.
Had to spend a lot of time adding some new functionality for this to work without taking up to much of the screen and without being too downscaled and being readable.
Finally done and happy with it for now.
Doesn't take up too much of the screen but is still readable when in fullscreen I think so yea.

Also removed the move button prompt from the enum and therefore from the button prompt assets since that is very straightforward and should be expected for a first person game and would also require creating more assets for the WASD being combined which is just unneeded.

Added new functionality in the main menu for disabling a new blocking volume in the level when entering the main menu. This blocking volume stops the player from going backwards into the loop collision without first going to the main menu since we want the player to follow the cat to the main menu and the bar itself.

Rest of the UI can be implemented whenever uploaded to the google drive now only missing a couple of buttons and stuff like turn indicator so yea.