Biel Model added. Placeholders for remaining UI thats not been uploaded. Update in game function in NPC interface. Fixed quit animation for pause menu. Changed main menu mesh.
Biel's model added.
The placeholder UI assets for darts have been added with just text inplace of the finished UI.

New update in game function in NPC interface for updating the bool in game? used for the animation blueprint.

All start game and end game function in game managers now update the in game bool for all NPCs playing the game so that their state in the animation blueprint will update accordingly.

Changed the main menu mesh to be a bin instead of a bin bag.

Fixed the quit widget animation wipe for the pause menu since it was not reversed and didn't realise until now.