BUG FIXING AFTER FINDING LOTS TRYING TO RECORD GAMEPLAY FOR COMPETITION. Darts end game fix. Auction unique card display fix. Auction actually checking if game over now. More checks in current hand component for hovering and selecting. Auction famine hovering and selected cards clearing. Changed tint of I doubt and Auction widgets. Card counters update more frequently for I Doubt It. Unique card indicator and doubt widget aesthetic updates in IDI and Auction. Clearing hovered and selected cards when adding cards for doubt. Changed doubt call widget location calculation. Changed played cards IDI location calculation. Think fixed bug with jukebox.
Darts actually uses the end game function in the game mode instead of using the leave function in the darts game manager.

Auction unique card display only actually shows if a unique card has been played.

Auction actually checking if the game is over now each round.

More checks in the current hand component in the hover check and select card function to see if the card that is attempting to be hovered or selected is actually in the current hand array and if not then we either unselect or hover the first card depending on the function this is in.

Auction famine now clears hovered and selected cards for the swapping current hand players.

Changed tints of a lot of widget components mainly in the IDI and auction game managers.

The card counters actually update for doubt in I doubt it now as well since they didn't before until the player played another card.

Doubt call and played cards locations now dependent on intended NPC locations rather than current actor locations for npc / players. And other widgets.

Think there was a bug with the jukebox not getting the player reference in the begin play and then failing to do the rest of the start song functionality so added a 0.2 second delay now to ensure the player is loaded. Also jukebox now actually starting on one of the 3 songs instead of one of the demo songs.